refactor(shared): withAuth HOC를 features/auth/lib로, 공통 UI 컴포넌트를 shared/ui로 이동
