+++ src/features/goods/ui/OptionSelectorItem/index.tsx
@@ -1,5 +1,5 @@
 import { GoodDetailOption } from "@/entities/goods/goods.types";
-import { ArrowBottomIcon } from "@/features/goods/ui/Icon";
+import { ArrowBottomIcon } from "@/shared/ui/Icon";
 import { formatNumberWithCommas } from "@/shared/utils/format";
 import * as s from "./style.css";
