+++ src/shared/ui/Spinner/style.css.ts
