+++ src/shared/ui/VisibilityLoader/style.css.ts
