+++ src/views/navigation/ui/NavigationBar/style.css.ts
