+++ src/views/navigation/ui/SearchForm/style.css.ts
